requirements/base.txt
Django==2.0.2
boto3==1.5.22
django-environ==0.4.4
django-model-utils==3.1.1
djangorestframework==3.7.7
mysqlclient==1.3.12
requests==2.18.4
warrant==0.6.1
Unipath==1.1
